+++ data-raw/prepare-sample-data.R
@@ -3,12 +3,14 @@
 
 devtools::load_all(".")
 simulated <- simulate_eventsurvey(seed = 10)
-sample_data <- data.frame(
-  day = simulated$day,
-  y = as.integer(cut(
-    simulated$y,
-    breaks = c(-Inf, 45, 48, 51, 54, Inf),
-    labels = FALSE
-  ))
-)
+sample_data <- simulated |>
+  dplyr::transmute(
+    day,
+    y = as.integer(cut(
+      y,
+      breaks = c(-Inf, 45, 48, 51, 54, Inf),
+      labels = FALSE
+    ))
+  ) |>
+  as.data.frame()
 save(sample_data, file = "data/sample_data.rda", compress = "xz")